+++ CHANGELOG.rst
@@ -28,6 +28,8 @@ New Features
 - Added GitHub action checking import time (:gh:`680` by `Julian Tachella`_)
 - Client model for server-side inference for using models in the cloud (:gh:`691` by `Andrew Wang`_)
 - Add dinv.utils.disable_tex to disable LaTeX (:gh:`726` by `Andrew Wang`_)
+- O(1) memory backprop for linear solver and example (:gh:`739` by `Minh Hai Nguyen`_)
+
 
 Changed
 ^^^^^^^